read_bills/rechnungsabrechnung_app/requirements.txt
streamlit>=1.28.0
pillow>=9.0.0
pytesseract>=0.3.10
python-dotenv>=0.19.0
groq>=0.4.0
pandas>=1.5.0
pdf2image>=1.16.0

